% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbricks-package.R
\docType{package}
\name{bbricks-package}
\alias{bbricks}
\alias{bbricks-package}
\title{bbricks: bbricks}
\description{
Basic building blocks in Bayesian network modeling.
}
\author{
\strong{Maintainer}: Haotian Chen \email{chenhaotian.jtt@gmail.com} (\href{https://orcid.org/https://orcid.org/0000-0001-9751-2093}{ORCID})

}
\keyword{internal}
