% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_Bricks.r
\name{.det}
\alias{.det}
\title{determinant of a matrix}
\usage{
.det(m)
}
\arguments{
\item{m}{numeric or matrix. When m is a numeric vector, it's treated as a diagonal matrix.}
}
\value{
logical, return TRUE if 'class2' is in 'class(object)'.
}
\description{
determinant of a matrix
}
