% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_Bricks.r
\name{.is}
\alias{.is}
\title{a internal version of "is", only for this package}
\usage{
.is(object, class2)
}
\arguments{
\item{object}{an object to be test with}

\item{class2}{the class you want to test}
}
\value{
logical, return TRUE if 'class2' is in 'class(object)'.
}
\description{
a internal version of "is", only for this package
}
