% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_Bricks.r
\name{.lmvgamma}
\alias{.lmvgamma}
\title{log multivariate gamma function}
\usage{
.lmvgamma(a, p)
}
\arguments{
\item{a}{numeric, a numeric vector to be evaluated.}

\item{p}{numeric, the dimension.}
}
\value{
numeric, the log value of the multivariate gamma function.
}
\description{
log multivariate gamma function
}
