% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_Inference.r
\name{dPosterior.GaussianGaussian}
\alias{dPosterior.GaussianGaussian}
\title{Density function of the posterior distribution of a "GaussianGaussian" object}
\usage{
\method{dPosterior}{GaussianGaussian}(obj, mu, LOG = TRUE, ...)
}
\arguments{
\item{obj}{A "GaussianGaussian" object.}

\item{mu}{matrix, or the ones that can be converted to matrix. Each row of mu is an sample.}

\item{LOG}{Return the log density if set to "TRUE".}

\item{...}{Additional arguments to be passed to other inherited types.}
}
\value{
A numeric vector of the same length as nrow(mu), the posterior density.
}
\description{
Generate the the density value of the posterior distribution of the following structure:
     \deqn{x ~ Gaussian(mu,Sigma)}
     \deqn{mu ~ Gaussian(m,S)}
Where Sigma is known. Gaussian() is the Gaussian distribution. See \code{?dGaussian} for the definition of Gaussian distribution.\cr
The model structure and prior parameters are stored in a "GaussianGaussian" object. \cr
Posterior density is the density function of Gaussian(mu|m,S).
}
\examples{
obj <- GaussianGaussian(gamma=list(Sigma=matrix(c(2,1,1,2),2,2),m=c(0.2,0.5),S=diag(2)))
mu <- rGaussian(10,mu=c(0,0),Sigma = diag(2))
dPosterior(obj=obj,mu=mu)
}
\seealso{
\code{\link{GaussianGaussian}}, \code{\link{rPosterior.GaussianGaussian}}
}
