% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.r
\docType{data}
\name{farmadsData}
\alias{farmadsData}
\title{farm ads data}
\format{A list of two elements:
\describe{
  \item{word}{character, the words}
  \item{document}{integer, document id of each word}
}}
\source{
\href{https://archive.ics.uci.edu/ml/datasets/Farm+Ads}{Farm-Ads}
}
\usage{
data(farmadsData)
}
\description{
A subset of farm ads data from https://archive.ics.uci.edu/ml/datasets/Farm+Ads
}
\keyword{datasets}
