% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet_Process.r
\name{\%plus\%}
\alias{\%plus\%}
\title{a plus b with NA values}
\usage{
e1 \%plus\% e2
}
\arguments{
\item{e1}{numeric}

\item{e2}{numeric}
}
\description{
make NA+NA=0 and NA+nonNA=0 instead of NA+NA=NA and NA+nonNA=NA
}
