% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.r
\docType{data}
\name{hlrData}
\alias{hlrData}
\title{Samples from a hierarchical linear model}
\format{A list of three elements:
\describe{
  \item{mathScore}{numeric, the mathScore of each student}
  \item{socioeconomicStatus}{numeric, the socioeconomic status score of each student}
  \item{schoolID}{integer, the school ID of each student}
}}
\usage{
data(lrData)
}
\description{
The data was part of the 2002 Educational Longitudinal Study (ELS), a survey of students from a large sample of schools across the United States. This dataset includes a population of schools as well as a population of students within each school.
}
\references{
Hoff, Peter D. A first course in Bayesian statistical methods. Vol. 580. New York: Springer, 2009.
}
\keyword{datasets}
