% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.r
\docType{data}
\name{lrData}
\alias{lrData}
\title{Samples from a simple linear model}
\format{A list of two elements:
\describe{
  \item{x}{numeric, linear samples}
  \item{X}{matrix, the "locations" of the linear samples}
}}
\usage{
data(lrData)
}
\description{
A list of two elements, a matrix "X" and a numeric vector "x". They came from a linear model: x = X*0.3 + epsilon, where epsilon is Gaussian distributed with mean 0 and variance 25.
}
\keyword{datasets}
