% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet_Process.r
\name{MAP.CatHDP2}
\alias{MAP.CatHDP2}
\title{MAP estimate for CatHDP}
\usage{
\method{MAP}{CatHDP2}(obj, ...)
}
\arguments{
\item{obj}{an object.}

\item{...}{Additional arguments to be passed to other inherited types.}
}
\description{
MAP estimate for CatHDP
}
