% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet_Process.r
\name{MAP.DP}
\alias{MAP.DP}
\title{MAP estimate of a "DP" object}
\usage{
\method{MAP}{DP}(obj, ...)
}
\arguments{
\item{obj}{An object of type "DP".}

\item{...}{Additional arguments to be passed to other inherited types.}
}
\description{
MAP estimate of a "DP" object
}
