% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet_Process.r
\name{MAP.HDP}
\alias{MAP.HDP}
\title{MAP estimate for HDP}
\usage{
\method{MAP}{HDP}(obj, ...)
}
\arguments{
\item{obj}{an object.}

\item{...}{Additional arguments to be passed to other inherited types.}
}
\description{
MAP estimate for HDP
}
