% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet_Process.r
\name{MAP.HDP2}
\alias{MAP.HDP2}
\title{MAP estimate for HDP2}
\usage{
\method{MAP}{HDP2}(obj, ...)
}
\arguments{
\item{obj}{an object.}

\item{...}{Additional arguments to be passed to other inherited types.}
}
\description{
MAP estimate for HDP2
}
