% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet_Process.r
\name{marginalLikelihood.CatHDP2}
\alias{marginalLikelihood.CatHDP2}
\title{Marginallikelihood for CatHDP2}
\usage{
\method{marginalLikelihood}{CatHDP2}(obj, ...)
}
\arguments{
\item{obj}{an object.}

\item{...}{Additional arguments to be passed to other inherited types.}
}
\description{
Marginallikelihood for CatHDP2
}
