% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet_Process.r
\name{marginalLikelihood.DP}
\alias{marginalLikelihood.DP}
\title{Marginallikelihood for DP}
\usage{
\method{marginalLikelihood}{DP}(obj, ...)
}
\arguments{
\item{obj}{an object.}

\item{...}{Additional arguments to be passed to other inherited types.}
}
\description{
Marginallikelihood for DP
}
