% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet_Process.r
\name{marginalLikelihood.HDP}
\alias{marginalLikelihood.HDP}
\title{Marginallikelihood for HDP}
\usage{
\method{marginalLikelihood}{HDP}(obj, ...)
}
\arguments{
\item{obj}{an object.}

\item{...}{Additional arguments to be passed to other inherited types.}
}
\value{
numeric, the marginal likelihood.
}
\description{
Marginallikelihood for HDP
}
