% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_Bricks.r
\name{marginalLikelihood}
\alias{marginalLikelihood}
\title{Get the marginal likelihood of a "BayesianBrick" object}
\usage{
marginalLikelihood(obj, ...)
}
\arguments{
\item{obj}{A "BayesianBrick" object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
numeric, the marginal likelihood
}
\description{
This is a generatic function that will generate the marginal likelihood of a set of observations conditioned on a given "BayesianBrick" object. i.e. for the model structure: \cr
     theta|gamma ~ H(gamma) \cr
     X|theta ~ F(theta) \cr
get p(X|gamma)
}
\seealso{
\code{\link{marginalLikelihood.GaussianNIW}} for Gaussian-NIW conjugate structure, \code{\link{marginalLikelihood.GaussianNIG}} for Gaussian-NIG conjugate structure, \code{\link{marginalLikelihood.CatDirichlet}} for Categorica-Dirichlet conjugate structure, \code{\link{marginalLikelihood.CatDP}} for Categorica-DP conjugate structure ...
}
