% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.r
\docType{data}
\name{mmData}
\alias{mmData}
\title{Samples from a mixture model}
\format{A matrix of Gaussian samples, each row is one sample.}
\usage{
data(mmData)
}
\description{
A matrix of 2-dimensional Gaussian mixture samples, the samples came from 4 different Gaussian distributions
}
\keyword{datasets}
