% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.r
\docType{data}
\name{mmhData}
\alias{mmhData}
\title{Samples from a hierarchical mixture model}
\format{A list of two elements:
\describe{
  \item{x}{matrix, 2-dimensional Gaussian samples, each row is a sample}
  \item{groupLabel}{integer, the group label of each Gaussian sample}
}}
\usage{
data(mmhData)
}
\description{
A list of 2-dimensional Gaussian mixture samples and the corresponding group labels. The samples are generated by 4 different Gaussian distributions. The samples are separated into 30 groups.
}
\keyword{datasets}
