% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.r
\docType{data}
\name{mmhhData}
\alias{mmhhData}
\title{Samples from a hierarchical mixture model with two layers of hierarchies}
\format{A list of two elements:
\describe{
  \item{x}{matrix, 2-dimensional Gaussian samples, each row is a sample}
  \item{groupLabel}{integer, the group label of each Gaussian sample}
  \item{subGroupLabel}{integer, the subgroup label of each Gaussian sample}
}}
\usage{
data(mmhhData)
}
\description{
A list of 2-dimensional Gaussian mixture samples and the corresponding group labels. The samples are generated by 4 different Gaussian distributions. Each sample is assigned to a group and a subgroup. There are 2 groups, group 1 has 10 subgroups, group 2 has 20 subgroups.
}
\keyword{datasets}
