% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet_Process.r
\name{MPE.DP}
\alias{MPE.DP}
\title{MPE estimate of a "DP" object}
\usage{
\method{MPE}{DP}(obj, ...)
}
\arguments{
\item{obj}{An object of type "DP".}

\item{...}{Additional arguments to be passed to other inherited types.}
}
\description{
MPE estimate of a "DP" object
}
