% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Categorical_Inference.r
\name{posterior.CatDirichlet}
\alias{posterior.CatDirichlet}
\title{Update a "CatDirichlet" object with sample sufficient statistics}
\usage{
\method{posterior}{CatDirichlet}(obj, ss, w = NULL, ...)
}
\arguments{
\item{obj}{A "CatDirichlet" object.}

\item{ss}{Sufficient statistics of x. In Categorical-Dirichlet case the sufficient statistic of sample x can be either x itself, of an "ssCat" object generated by the function sufficientStatistics.CatDirichlet().}

\item{w}{Sample weights, default NULL.}

\item{...}{Additional arguments to be passed to other inherited types.}
}
\value{
None. the gamma stored in "obj" will be updated based on "ss".
}
\description{
For the model structure: \cr
     pi|alpha ~ Dirichlet(alpha) \cr
     x|pi ~ Categorical(pi) \cr
update alpha by adding the information of newly observed samples x. \cr
The model structure and prior parameters are stored in a "CatDirichlet" object, the prior parameters in this object will be updated after running this function.
}
\examples{
obj <- CatDirichlet(gamma=list(alpha=rep(1,26),uniqueLabels = letters))
x <- sample(letters,size = 20,replace = TRUE)
w <- runif(20)
posterior(obj=obj,ss=x)
obj
posteriorDiscard(obj=obj,ss=x)
obj
## weighted sample
posterior(obj=obj,ss=x,w=w)
obj
posteriorDiscard(obj=obj,ss=x,w=w)
obj
}
\references{
Murphy, Kevin P. Machine learning: a probabilistic perspective. MIT press, 2012.
}
\seealso{
\code{\link{CatDirichlet}},\code{\link{posteriorDiscard.CatDirichlet}}
}
