% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Categorical_Inference.r
\name{posterior_bySufficientStatistics.CatDirichlet}
\alias{posterior_bySufficientStatistics.CatDirichlet}
\title{Update a "CatDirichlet" object with sample sufficient statistics}
\usage{
\method{posterior_bySufficientStatistics}{CatDirichlet}(obj, ss)
}
\description{
Update a "CatDirichlet" object with sample sufficient statistics
}
\examples{
obj <- CatDirichlet(gamma=list(alpha=rep(1,26),uniqueLabels = letters))
x <- sample(letters,size = 20,replace = TRUE)
w <- runif(20)
posterior(obj=obj,ss=x)
obj
posteriorDiscard(obj=obj,ss=x)
obj
## by sufficient statistics
ss <- sufficientStatistics(obj = obj,x=x)
posterior_bySufficientStatistics(obj = obj,ss=ss)
obj
## weighted sample
posterior(obj=obj,ss=x,w=w)
obj$gamma$a
posteriorDiscard(obj=obj,ss=x,w=w)
obj$gamma$a
## by sufficient statistics
posterior_bySufficientStatistics(obj = obj,ss=ss)
obj$gamma$a
posteriorDiscard_bySufficientStatistics(obj = obj,ss=ss)
obj$gamma$a
}
