% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet_Process.r
\name{posterior_bySufficientStatistics.CatHDP}
\alias{posterior_bySufficientStatistics.CatHDP}
\title{Update a "CatHDP" object with sample sufficient statistics}
\usage{
\method{posterior_bySufficientStatistics}{CatHDP}(obj, ss1, ss2, j)
}
\description{
Update a "CatHDP" object with sample sufficient statistics
}
