% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet_Process.r
\name{posterior_bySufficientStatistics.CatHDP2}
\alias{posterior_bySufficientStatistics.CatHDP2}
\title{Update a "CatHDP2" object with sample sufficient statistics}
\usage{
\method{posterior_bySufficientStatistics}{CatHDP2}(obj, ss1, ss2, ss3, m, j)
}
\description{
Update a "CatHDP2" object with sample sufficient statistics
}
