% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_Inference.r
\name{posterior_bySufficientStatistics.GaussianNIG}
\alias{posterior_bySufficientStatistics.GaussianNIG}
\title{Update a "GaussianNIG" object with sample sufficient statistics}
\usage{
\method{posterior_bySufficientStatistics}{GaussianNIG}(obj, ss, w = NULL)
}
\description{
Update a "GaussianNIG" object with sample sufficient statistics
}
