% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_Bricks.r
\name{posterior_bySufficientStatistics}
\alias{posterior_bySufficientStatistics}
\title{update the prior distribution with sufficient statistics}
\usage{
posterior_bySufficientStatistics(obj, ...)
}
\description{
update the prior distribution with sufficient statistics
}
