% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Categorical_Inference.r
\name{posteriorDiscard_bySufficientStatistics.CatDirichlet}
\alias{posteriorDiscard_bySufficientStatistics.CatDirichlet}
\title{Update the prior Dirichlet distribution with sample sufficient statistics}
\usage{
\method{posteriorDiscard_bySufficientStatistics}{CatDirichlet}(obj, ss)
}
\description{
Update the prior Dirichlet distribution with sample sufficient statistics
}
\seealso{
\code{\link{posteriorDiscard}}
}
