% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet_Process.r
\name{posteriorDiscard_bySufficientStatistics.CatDP}
\alias{posteriorDiscard_bySufficientStatistics.CatDP}
\title{Update a "CatDP" object with sample sufficient statistics}
\usage{
\method{posteriorDiscard_bySufficientStatistics}{CatDP}(obj, ss)
}
\description{
Update a "CatDP" object with sample sufficient statistics
}
