% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_Inference.r
\name{posteriorDiscard_bySufficientStatistics.GaussianNIG}
\alias{posteriorDiscard_bySufficientStatistics.GaussianNIG}
\title{Update a "GaussianNIG" object with sample sufficient statistics}
\usage{
\method{posteriorDiscard_bySufficientStatistics}{GaussianNIG}(obj, ss, w = NULL, ...)
}
\arguments{
\item{obj}{A "GaussianNIG" object.}

\item{ss}{Sufficient statistics of (x,X). In Gaussian-NIG case the sufficient statistic of sample (x,X) is a object of type "ssGaussianLinear", it can be  generated by the function sufficientStatistics().}

\item{w}{Sample weights,default NULL.}

\item{...}{Additional arguments to be passed to other inherited types.}
}
\description{
Update a "GaussianNIG" object with sample sufficient statistics
}
