% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_Inference.r
\name{posteriorDiscard_bySufficientStatistics.GaussianNIW}
\alias{posteriorDiscard_bySufficientStatistics.GaussianNIW}
\title{Update a "GaussianNIW" object with sample sufficient statistics}
\usage{
\method{posteriorDiscard_bySufficientStatistics}{GaussianNIW}(obj, ss, w = NULL, ...)
}
\arguments{
\item{obj}{A "GaussianNIW" object.}

\item{ss}{Sufficient statistics of x. In Gaussian-NIW case the sufficient statistic of sample x is a object of type "ssGaussian", it can be  generated by the function sufficientStatistics().}

\item{w}{Sample weights,default NULL.}

\item{...}{Additional arguments to be passed to other inherited types.}
}
\description{
Update a "GaussianNIW" object with sample sufficient statistics
}
