% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_Bricks.r
\name{posteriorDiscard_bySufficientStatistics}
\alias{posteriorDiscard_bySufficientStatistics}
\title{update the prior distribution with sufficient statistics}
\usage{
posteriorDiscard_bySufficientStatistics(obj, ...)
}
\description{
update the prior distribution with sufficient statistics
}
