% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_Bricks.r
\name{print.BasicBayesian}
\alias{print.BasicBayesian}
\title{Print the content of an BayesicBasyeisn object}
\usage{
\method{print}{BasicBayesian}(obj)
}
\description{
Print the content of an BayesicBasyeisn object
}
