% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet_Process.r
\name{print.CatHDP}
\alias{print.CatHDP}
\title{Print the content of an CatHDP object}
\usage{
\method{print}{CatHDP}(x, ...)
}
\arguments{
\item{x}{An object of type "CatHDP".}

\item{...}{Additional arguments to be passed to other inherited types.}
}
\value{
None.
}
\description{
Print the content of an CatHDP object
}
