% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet_Process.r
\name{print.CatHDP2}
\alias{print.CatHDP2}
\title{Print the content of an CatHDP2 object}
\usage{
\method{print}{CatHDP2}(x, ...)
}
\arguments{
\item{x}{An object of type "CatHDP2".}

\item{...}{Additional arguments to be passed to other inherited types.}
}
\value{
None.
}
\description{
Print the content of an CatHDP2 object
}
