% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet_Process.r
\name{print.DP}
\alias{print.DP}
\title{print the content of a "DP" object}
\usage{
\method{print}{DP}(x, ...)
}
\arguments{
\item{x}{An object of type "DP".}

\item{...}{Additional arguments to be passed to other inherited types.}
}
\value{
None.
}
\description{
print the content of a "DP" object
}
