% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet_Process.r
\name{print.HDP2}
\alias{print.HDP2}
\title{print the content of a "HDP2" object}
\usage{
\method{print}{HDP2}(x, ...)
}
\arguments{
\item{x}{An object of type "HDP2".}

\item{...}{Additional arguments to be passed to other inherited types.}
}
\value{
None.
}
\description{
print the content of a "HDP2" object
}
