% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Categorical_Inference.r
\name{rCategorical}
\alias{rCategorical}
\title{Random generation for Categorical distribution}
\usage{
rCategorical(n, p)
}
\arguments{
\item{n}{integer, number of samples.}

\item{p}{numeric, probabilities. length(p)=K.}
}
\value{
An integer vector of length n.
}
\description{
Generate random integer samples from a Categorical distribution. For a random variable x, the density function of categorical distribution is defined as
  \deqn{prod_{k in 1:K} p_k^{I(x=k)}}
Where K is the number of unique values.
}
\examples{
rCategorical(n=20,p=c(1,2))
}
\seealso{
\code{\link{dCategorical}}
}
