% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Categorical_Inference.r
\name{rDir}
\alias{rDir}
\title{Random generation for Dirichelt distribution}
\usage{
rDir(n, alpha)
}
\arguments{
\item{n}{integer, number of samples.}

\item{alpha}{numeric, Dirichlet parameter.}
}
\value{
A numeric matrix of n rows and length(alpha) columns.
}
\description{
Generate random samples from Dirichlet distribution.
}
\examples{
rDir(5,c(1,2,3)) #generate 5 samples with parameters c(1,2,3)
}
\seealso{
\code{\link{dDir}}
}
