% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_Bricks.r
\name{release_questions}
\alias{release_questions}
\title{additional release questions}
\usage{
release_questions()
}
\description{
This is a list of additional questions you want devtools::release() to ask when releasing your package.
}
