% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_Bricks.r
\name{rObservationDistribution}
\alias{rObservationDistribution}
\title{random number generation of the observation distribution}
\usage{
rObservationDistribution(obj, ...)
}
\arguments{
\item{obj}{An object of the target type.}

\item{...}{other parameters.}
}
\description{
random number generation of the observation distribution
}
