% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet_Process.r
\name{sufficientStatistics.CatHDP}
\alias{sufficientStatistics.CatHDP}
\title{Sufficient statistics of a "CatHDP" object}
\usage{
\method{sufficientStatistics}{CatHDP}(obj, z, k, foreach = FALSE)
}
\arguments{
\item{obj}{A "CatHDP" object.}

\item{z}{integer, the elements of the vector must all greater than 0, the samples of a Categorical distribution.}

\item{k}{integer, the elements of the vector must all greater than 0, the samples of a Categorical distribution.}

\item{foreach}{logical, specifying whether to return the sufficient staistics for each observation. Default FALSE.}
}
\value{
An object of class "ssCatHDP", the sufficient statistics of a set of categorical samples. Or an integer matrix if foreach = TRUE.
}
\description{
For following model structure: \cr
     G_j|gamma ~ DP(gamma,U), j = 1:J \cr
     pi_j|G_j,alpha ~ DP(alpha,G_j) \cr
     z|pi_j ~ Categorical(pi_j) \cr
     k|z,G_j ~ Categorical(G_j), if z is a sample from the base measure G \cr
where DP(gamma,U) is a Dirichlet Process on integers, gamma is the concentration parameter of the Dirichlet Process. DP(gamma,G_j) is a Dirichlet Process on integers with concentration parameter alpha and base measure G_j. \cr
In the case of CatHDP, z and k can only be positive integers. \cr
The model structure and prior parameters are stored in a "CatHDP" object. \cr
The sufficient statistics of a set of samples z and k are the effective counts of the unique positive integers.
}
\seealso{
\code{\link{CatHDP}}, \code{\link{sufficientStatistics_Weighted.CatHDP}}
}
