% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet_Process.r
\name{sufficientStatistics.CatHDP2}
\alias{sufficientStatistics.CatHDP2}
\title{Sufficient statistics of a "CatHDP2" object}
\usage{
\method{sufficientStatistics}{CatHDP2}(obj, x, foreach = FALSE)
}
\description{
Sufficient statistics of a "CatHDP2" object
}
