% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_Inference.r
\name{sufficientStatisticsGaussian_Weighted}
\alias{sufficientStatisticsGaussian_Weighted}
\title{Sufficient statistics for Gaussian samples}
\usage{
sufficientStatisticsGaussian_Weighted(x, w, foreach = FALSE)
}
\arguments{
\item{x, }{matrix, Gaussian samples, when x is a matrix, each row is an sample of dimension ncol(x). when x is a vector, x is length(x) samples of dimension 1}

\item{w}{numeric, sample weights}

\item{foreach}{logical, if foreach=TRUE, will return a list of sufficient statistics for each row of x, otherwise will return the sufficent statistics of x as a whole}
}
\value{
if foreach=TRUE, will return a list of sufficient statistics for each row of x, otherwise will return the sufficent statistics of x as a whole
}
\description{
for a design matrix x(each row is a samples), the sufficient statistics are \cr
 the effective number of samples N=nrow(x) \cr
 the sample sum xsum = colSums(x) \cr
 the uncentered scatter matrix S = t(x)%*%x
}
\examples{
\dontrun{
x <- rGaussian(10,mu = c(-1.5,1.5),Sigma = matrix(c(0.1,0.03,0.03,0.1),2,2))
w <- runif(10)
sufficientStatisticsGaussian_Weighted(x,w,foreach = FALSE)
sufficientStatisticsGaussian(x,foreach = TRUE)
}
}
