% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_Bricks.r
\name{BasicBayesian}
\alias{BasicBayesian}
\title{Create objects of type '"BasicBayesian"'.}
\usage{
BasicBayesian(ENV = parent.frame())
}
\arguments{
\item{ENV}{The environment where you want to create the BasicBayesian object}
}
\value{
An object of class "BasicBayesian"
}
\description{
A Basic Bayesian Object is with following conditional dependency structure: \cr
     theta|gamma ~ H(gamma) \cr
     X|theta ~ F(theta) \cr
Where H(gamma) is usually called "the prior distribution", F(theta) is called "the observation distribution". Objects of type "GaussianNIW", "GaussianNIG", "CatDirichlet", and "CatDP" are all "BasicBayesian"s.
}
\seealso{
\code{\link{GaussianNIW}} for Gaussian-NIW conjugate structure, \code{\link{GaussianNIG}} for Gaussian-NIG conjugate structure, \code{\link{CatDirichlet}} for Categorical-Dirichlet conjugate structure, \code{\link{CatDP}} for Categorical-DP conjugate structure ...
}
