% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbricks-package.R
\docType{package}
\name{bbricks-package}
\alias{bbricks}
\alias{bbricks-package}
\title{bbricks: Bayesian Bricks}
\description{
Basic building blocks in Bayesian modeling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chenhaotian/Bayesian-Bricks}
  \item Report bugs at \url{https://github.com/chenhaotian/Bayesian-Bricks/issues}
}

}
\author{
\strong{Maintainer}: Haotian Chen \email{chenhaotian.jtt@gmail.com} (\href{https://orcid.org/0000-0001-9751-2093}{ORCID})

}
\keyword{internal}
