% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbricks-package.R
\docType{package}
\name{bbricks-package}
\alias{bbricks}
\alias{bbricks-package}
\title{bbricks: Bayesian methods and graphical model structures for statistical modeling}
\description{
A class of frequently used Bayesian parametric and nonparametric model structures and the corresponding inference tasks. Structures include Gaussian and Normal-Inverse-Wishart conjugate structure, Gaussian and Normal-Inverse-Gamma conjugate structure, Categorical and Dirichlet conjugate structure, Dirichlet Process on positive integers, Dirichlet Process in general, Hierarchical Dirichlet Process and so on; Tasks include updating posteriors, calculating marginal likelihood, calculating posterior predictives, calculating MAP estimates and so on.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chenhaotian/Bayesian-Bricks}
  \item Report bugs at \url{https://github.com/chenhaotian/Bayesian-Bricks/issues}
}

}
\author{
\strong{Maintainer}: Haotian Chen \email{chenhaotian.jtt@gmail.com} (\href{https://orcid.org/0000-0001-9751-2093}{ORCID})

}
\keyword{internal}
