% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbricks-package.R
\docType{package}
\name{bbricks-package}
\alias{bbricks}
\alias{bbricks-package}
\title{bbricks: Bayesian Methods and Graphical Model Structures for Statistical Modeling}
\description{
A set of frequently used Bayesian parametric and nonparametric model structures, as well as a set of tools for common analytical tasks. Structures include linear Gaussian systems, Gaussian and Normal-Inverse-Wishart conjugate structure, Gaussian and Normal-Inverse-Gamma conjugate structure, Categorical and Dirichlet conjugate structure, Dirichlet Process on positive integers, Dirichlet Process in general, Hierarchical Dirichlet Process ... Tasks include updating posteriors, sampling from posteriors, calculating marginal likelihood, calculating posterior predictive densities, sampling from posterior predictive distributions, calculating "Maximum A Posteriori" (MAP) estimates ... See <https://chenhaotian.github.io/Bayesian-Bricks/> to get started.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chenhaotian/Bayesian-Bricks}
  \item Report bugs at \url{https://github.com/chenhaotian/Bayesian-Bricks/issues}
}

}
\author{
\strong{Maintainer}: Haotian Chen \email{chenhaotian.jtt@gmail.com} (\href{https://orcid.org/0000-0001-9751-2093}{ORCID})

}
\keyword{internal}
