% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Categorical_Inference.r
\name{dCategorical}
\alias{dCategorical}
\title{Probability mass function for Categorical distribution}
\usage{
dCategorical(x, p)
}
\arguments{
\item{x}{integer, categorical samples.}

\item{p}{numeric, probabilities.}
}
\value{
A numeric vector of the same length of 'x'.
}
\description{
Calculate probability masses for integer valued Categorical random samples.
For a random variable x, the density function of categorical distribution is defined as
  \deqn{prod_{k in 1:K} p_k^{I(x=k)}}
Where K is the number of unique values.
}
\examples{
\donttest{
dCategorical(x=c(1L,2L,1L),p=c(1,2))
}
}
\seealso{
\code{\link{rCategorical}}
}
