% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_Bricks.r
\name{dObservationDistribution}
\alias{dObservationDistribution}
\title{the density of observation distribution}
\usage{
dObservationDistribution(obj, ...)
}
\description{
the density of observation distribution
}
