% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet_Process.r
\name{\%plus\%}
\alias{\%plus\%}
\title{a plus b with NA values}
\usage{
e1 \%plus\% e2
}
\arguments{
\item{e1}{numeric}

\item{e2}{numeric}
}
\value{
numeric, the sum of "e1" and "e2". The length of the returned vector is the same length as the longest of "e1" or "e2".
}
\description{
make NA+NA=0 and NA+(not NA)=0 instead of NA+NA=NA and NA+(not NA)=NA
}
