% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_Inference.r
\name{MPE.GaussianNIW}
\alias{MPE.GaussianNIW}
\title{MPE of a "GaussianNIW" object}
\usage{
\method{MPE}{GaussianNIW}(obj)
}
\arguments{
\item{obj}{A "GaussianNIW" object.}
}
\value{
A named list, the MPE estimate of "theta".
}
\description{
Generate the MPE of "theta" in following GaussianNIW structure: \cr
     theta|gamma ~ NIW(gamma) \cr
     x|theta ~ Gaussian(theta) \cr
where theta = (mu,Sigma) is the Gaussian parameter, gamma = (m,k,v,S) is the Normal-Inverse-Wishart(NIW) parameter. \cr
The model structure and prior parameters are stored in a "GaussianNIW" object. \cr
MPE is theta_MPE = mean_theta(theta|gamma,x)
}
\examples{
## update the piror with new observations then calculate the MAP estimate
x <- rGaussian(1000,mu = c(1,1),Sigma = matrix(c(1,0.5,0.5,3),2,2))
obj <- GaussianNIW(gamma=list(m=c(0,0),k=1,v=2,S=diag(2)))
ss <- sufficientStatistics_Weighted(obj = obj,x=x,w=w,foreach = TRUE)
for(i in 1L:length(ss)) posterior(obj = obj,ss=ss[[i]])
MPE(obj)
}
\seealso{
\code{\link{GaussianNIW}}
}
