% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet_Process.r
\name{posterior_bySufficientStatistics.CatDP}
\alias{posterior_bySufficientStatistics.CatDP}
\title{Update a "CatDP" object with sample sufficient statistics}
\usage{
\method{posterior_bySufficientStatistics}{CatDP}(obj, ss, ...)
}
\arguments{
\item{obj}{A "CatDP" object.}

\item{ss}{Sufficient statistics of x. In Categorical-DP case the sufficient statistic of sample x can either be an object of type "ssCatDP" generated by sufficientStatistics(), or x itself(if x is a integer vector with all positive values).}

\item{...}{Additional arguments to be passed to other inherited types.}
}
\value{
None. the model stored in "obj" will be updated based on "ss".
}
\description{
Update a "CatDP" object with sample sufficient statistics
}
\references{
Teh, Yee W., et al. "Sharing clusters among related groups: Hierarchical Dirichlet processes." Advances in neural information processing systems. 2005.
}
