% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_Inference.r
\name{posterior_bySufficientStatistics.GaussianNIW}
\alias{posterior_bySufficientStatistics.GaussianNIW}
\title{Update a "GaussianNIW" object with sample sufficient statistics}
\usage{
\method{posterior_bySufficientStatistics}{GaussianNIW}(obj, ss, w = NULL)
}
\description{
Update a "GaussianNIW" object with sample sufficient statistics
}
