% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_Bricks.r
\name{posterior_bySufficientStatistics}
\alias{posterior_bySufficientStatistics}
\title{update the prior distribution with sufficient statistics}
\usage{
posterior_bySufficientStatistics(obj, ...)
}
\arguments{
\item{obj}{A "BayesianBrick" object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
None, or an error message if the update fails.
}
\description{
update the prior distribution with sufficient statistics
}
