% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Categorical_Inference.r
\name{posteriorDiscard_bySufficientStatistics.CatDirichlet}
\alias{posteriorDiscard_bySufficientStatistics.CatDirichlet}
\title{Update the prior Dirichlet distribution with sample sufficient statistics}
\usage{
\method{posteriorDiscard_bySufficientStatistics}{CatDirichlet}(obj, ss, ...)
}
\arguments{
\item{obj}{A "CatDirichlet" object.}

\item{ss}{Sufficient statistics of x. In Categorical-Dirichlet case the sufficient statistic of sample x can be either x itself, of an "ssCat" object generated by the function sufficientStatistics.CatDirichlet().}

\item{...}{Additional arguments to be passed to other inherited types.}
}
\value{
None. the gamma stored in "obj" will be updated based on "ss".
}
\description{
Update the prior Dirichlet distribution with sample sufficient statistics
}
\references{
Murphy, Kevin P. Machine learning: a probabilistic perspective. MIT press, 2012.
}
\seealso{
\code{\link{posteriorDiscard}}
}
