% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet_Process.r
\name{posteriorDiscard_bySufficientStatistics.CatHDP}
\alias{posteriorDiscard_bySufficientStatistics.CatHDP}
\title{Update a "CatHDP" object with sample sufficient statistics}
\usage{
\method{posteriorDiscard_bySufficientStatistics}{CatHDP}(obj, ss1, ss2, j)
}
\description{
Update a "CatHDP" object with sample sufficient statistics
}
