% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_Inference.r
\name{posteriorDiscard_bySufficientStatistics.GaussianNIW}
\alias{posteriorDiscard_bySufficientStatistics.GaussianNIW}
\title{Update a "GaussianNIW" object with sample sufficient statistics}
\usage{
\method{posteriorDiscard_bySufficientStatistics}{GaussianNIW}(obj, ss, w = NULL)
}
\description{
Update a "GaussianNIW" object with sample sufficient statistics
}
