% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_Bricks.r
\name{posteriorDiscard_bySufficientStatistics}
\alias{posteriorDiscard_bySufficientStatistics}
\title{update the prior distribution with sufficient statistics}
\usage{
posteriorDiscard_bySufficientStatistics(obj, ...)
}
\arguments{
\item{obj}{A "BayesianBrick" object used to select a method.}

\item{...}{other parameters.}
}
\value{
None, or an error message if the update fails.
}
\description{
update the prior distribution with sufficient statistics
}
