% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_Bricks.r
\name{print.BasicBayesian}
\alias{print.BasicBayesian}
\title{Print the content of an BasicBasyesian object}
\usage{
\method{print}{BasicBayesian}(x, ...)
}
\arguments{
\item{x}{An object of the target type.}

\item{...}{other parameters passed to print.}
}
\value{
None.
}
\description{
Print the content of an BasicBasyesian object
}
