% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Categorical_Inference.r
\name{rCategorical}
\alias{rCategorical}
\title{Random generation for Categorical distribution}
\usage{
rCategorical(n, p)
}
\arguments{
\item{n}{integer, number of samples.}

\item{p}{numeric, probabilities.}
}
\value{
An integer vector of length n.
}
\description{
Generate random integer samples from a Categorical distribution.
}
\examples{
rCategorical(n=20,p=c(1,2))
}
\seealso{
\code{\link{dCategorical}}
}
