% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet_Process.r
\name{sufficientStatistics_Weighted.CatHDP2}
\alias{sufficientStatistics_Weighted.CatHDP2}
\title{Sufficient statistics of a "CatHDP2" object}
\usage{
\method{sufficientStatistics_Weighted}{CatHDP2}(obj, x, w, foreach = FALSE)
}
\description{
Sufficient statistics of a "CatHDP2" object
}
